Update field mappings
